style(dir): move src and test files to separate folders
